defining some min max midi param values. Applying midiLearn to those

+++ farq2014/src/ParameterInputGen/DTMidiMap.h
@@ -21,8 +21,8 @@ public:
     string  paramId;
     int     inputMinValue;
     int     inputMaxValue;
-    int     paramMinValue;
-    int     paramMaxValue;
+    float     paramMinValue;
+    float     paramMaxValue;
 };
